`lsp-imenu': use `lsp-symbol-kinds'
Rename `lsp--symbol-kind' to `lsp-symbol-kinds', making the former an
alias for the latter and the latter a `defcustom'. In it, rename "Enum
Member" to "Enumerator" since that looks better.

Instead of creating another constant for that does the same as
`lsp-symbol-kinds', use the latter.

Refactor: simplify `lsp--get-symbol-type' to use `alist-get' instead of
a complex threading sequence, simplifying it to a single s-expression.